// src/common/enum/error.enum.ts
export enum ErrorEnum {
  CREATE_ERROR = 'E-0000',
  NOT_FOUND = 'E-0001',
  ALREADY_EXISTS = 'E-0002',
  NOT_AUTHORIZED = 'E-0003',
  INVALID_CREDENTIALS = 'E-0004',
  VALIDATION_ERROR = 'E-0005',
  TOO_MANY_REQUESTS = 'E-0006',
  RATE_LIMIT_EXCEEDED = 'E-0007',
  IP_BLOCKED = 'E-0008',
  BRUTE_FORCE_DETECTED = 'E-0009',
  ACCOUNT_LOCKED = 'E-0010',
  BAD_REQUEST = "E-0011",
}

export enum ErrorPrisma {
  P1000 = 'C-1000',
  P1001 = 'C-1001',
  P1002 = 'C-1002',
  P1003 = 'C-1003',
  P1008 = 'C-1008',
  P1009 = 'C-1009',
  P1010 = 'C-1010',
  P1011 = 'C-1011',
  P1012 = 'C-1012',
  P1013 = 'C-1013',
  P1014 = 'C-1014',
  P1015 = 'C-1015',
  P1016 = 'C-1016',
  P1017 = 'C-1017',
  P2000 = 'P-2000',
  P2001 = 'P-2000',
  P2002 = 'P-2000',
  P2003 = 'P-2000',
  P2004 = 'P-2000',
  P2005 = 'P-2000',
  P2006 = 'P-2000',
  P2007 = 'P-2000',
  P2008 = 'P-2000',
  P2009 = 'P-2000',
  P2010 = 'P-2000',
  P2011 = 'P-2000',
  P2012 = 'P-2000',
  P2013 = 'P-2000',
  P2014 = 'P-2000',
  P2015 = 'P-2000',
  P2016 = 'P-2000',
  P2017 = 'P-2000',
  P2018 = 'P-2000',
  P2019 = 'P-2000',
  P2020 = 'P-2000',
  P2021 = 'P-2000',
  P2022 = 'P-2000',
  P2023 = 'P-2000',
  P2024 = 'P-2000',
  P2025 = 'P-2000',
  P2026 = 'P-2000',
  P2027 = 'P-2000',
  P2028 = 'P-2000',
  P2029 = 'P-2000',
  P2030 = 'P-2000',
  P2031 = 'P-2000',
  P2033 = 'P-2000',
  P2034 = 'P-2000',
  P2035 = 'P-2000',
  P2036 = 'P-2000',
  P2037 = 'P-2000',
  P3000 = 'M-3000',
  P3001 = 'M-3001',
  P3002 = 'M-3002',
  P3003 = 'M-3003',
  P3004 = 'M-3004',
  P3005 = 'M-3005',
  P3006 = 'M-3006',
  P3007 = 'M-3007',
  P3008 = 'M-3008',
  P3009 = 'M-3009',
  P3010 = 'M-3010',
  P3011 = 'M-3011',
  P3012 = 'M-3012',
  P3013 = 'M-3013',
  P3014 = 'M-3014',
  P3015 = 'M-3015',
  P3016 = 'M-3016',
  P3017 = 'M-3017',
  P3018 = 'M-3018',
  P3019 = 'M-3019',
  P3020 = 'M-3020',
  P3021 = 'M-3021',
  P3022 = 'M-3022',
  P3023 = 'M-3023',
  P3024 = 'M-3024',
  P4000 = 'S-4000',
  P4001 = 'S-4001',
  P4002 = 'S-4002',
  P6000 = 'A-6000',
  P6001 = 'A-6001',
  P6002 = 'A-6002',
  P6003 = 'A-6003',
  P6004 = 'A-6004',
  P6005 = 'A-6005',
  P6006 = 'A-6006',
  P6008 = 'A-6008',
  P6009 = 'A-6009',
  P6010 = 'A-6010',
  P5011 = 'A-5011',
}